add recycler view adapter of main page
