+++ app/src/main/java/app/jerry/drink/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package app.jerry.drink
+package app.jerry.drink.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,20 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import app.jerry.drink.databinding.FragmentOrderBinding
-import app.jerry.drink.databinding.FragmentRadarBinding
+import app.jerry.drink.R
+import app.jerry.drink.databinding.FragmentHomeBinding
 
-class RadarFragment : Fragment() {
+class HomeFragment : Fragment() {
 
 
-    lateinit var binding: FragmentRadarBinding
+    lateinit var binding: FragmentHomeBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         binding = DataBindingUtil.inflate(
-            inflater, R.layout.fragment_radar, container, false
+            inflater, R.layout.fragment_home, container, false
         )
 
         return binding.root
