+++ app/src/main/java/app/jerry/drink/order/OrderFragment.kt
@@ -1,4 +1,4 @@
-package app.jerry.drink
+package app.jerry.drink.order
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import app.jerry.drink.databinding.FragmentHomeBinding
+import app.jerry.drink.R
 import app.jerry.drink.databinding.FragmentOrderBinding
 
 class OrderFragment : Fragment() {
