+++ app/src/main/java/app/jerry/drink/radar/RadarFragment.kt
@@ -1,4 +1,4 @@
-package app.jerry.drink
+package app.jerry.drink.radar
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,19 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import app.jerry.drink.databinding.FragmentHomeBinding
+import app.jerry.drink.R
+import app.jerry.drink.databinding.FragmentRadarBinding
 
-class HomeFragment : Fragment() {
+class RadarFragment : Fragment() {
 
 
-    lateinit var binding: FragmentHomeBinding
+    lateinit var binding: FragmentRadarBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         binding = DataBindingUtil.inflate(
-            inflater, R.layout.fragment_home, container, false
+            inflater, R.layout.fragment_radar, container, false
         )
 
         return binding.root
